backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.23.2
motor>=3.3.1
pydantic>=2.9.0
pydantic-settings>=2.0.3
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
email-validator>=2.0.0
